requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cachetools==5.5.0
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.8.0
PyYAML==6.0.2
sniffio==1.3.1
typing_extensions==4.12.2
zhipuai==2.1.5.20230904
